Add Task Appending to Array and DOM
This commit adds the functionality to add a new task to an array, as well as the DOM. It also sets up images for the checkmark, edit, and delete buttons. It adds checkmark functionality to mark the task as complete in the array and DOM if clicked, and adds the remove functionality to remove it from both the DOM and the array of tasks. Extra CSS styling was sprinkled in as well.